split out scripts from init.py
